+++ core/components/tinymcerte/src/TinyMCERTE.php
@@ -31,7 +31,7 @@ class TinyMCERTE
      * The version
      * @var string $version
      */
-    public $version = '2.0.5';
+    public $version = '2.0.6';
 
     /**
      * The class options