+++ core/components/tinymcerte/docs/changelog.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.6] - 2022-02-16
+
+### Added
+
+- Change `lightgray` skin system setting to `modx` during an update
+- Add `autoresize` plugin per default
+
 ## [2.0.5] - 2021-12-27
 
 ### Added